Refactor install code for both OpenCL and LevelZero
