+++ tornado-drivers/spirv/src/main/java/uk/ac/manchester/tornado/drivers/spirv/SPIRVCodeCache.java
@@ -2,7 +2,7 @@
  * This file is part of Tornado: A heterogeneous programming framework:
  * https://github.com/beehive-lab/tornadovm
  *
- * Copyright (c) 2021-2022 APT Group, Department of Computer Science,
+ * Copyright (c) 2021-2022, 2024, APT Group, Department of Computer Science,
  * School of Engineering, The University of Manchester. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
@@ -24,13 +24,18 @@
 package uk.ac.manchester.tornado.drivers.spirv;
 
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.nio.ByteBuffer;
+import java.nio.ByteOrder;
 import java.nio.file.FileSystems;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.concurrent.ConcurrentHashMap;
 
+import uk.ac.manchester.tornado.api.exceptions.TornadoBailoutRuntimeException;
 import uk.ac.manchester.tornado.drivers.spirv.graal.SPIRVInstalledCode;
+import uk.ac.manchester.tornado.runtime.common.TornadoOptions;
 import uk.ac.manchester.tornado.runtime.tasks.meta.TaskMetaData;
 
 public abstract class SPIRVCodeCache {
@@ -59,7 +64,7 @@ public abstract class SPIRVCodeCache {
     }
 
     public SPIRVInstalledCode getInstalledCode(String id, String entryPoint) {
-        return cache.get(STR."\{id}-\{entryPoint}");
+        return cache.get(id + "-" + entryPoint);
     }
 
     protected void writeBufferToFile(ByteBuffer buffer, String filepath) {
@@ -87,8 +92,30 @@ public abstract class SPIRVCodeCache {
         return tempDirectory + pathSeparator + user + pathSeparator + "tornadoVM-spirv";
     }
 
+    public SPIRVInstalledCode installSPIRVBinary(TaskMetaData meta, String id, String entryPoint, byte[] binary) {
+        if (binary == null || binary.length == 0) {
+            throw new RuntimeException("[ERROR] SPIR-V Binary Module is Empty");
+        }
+        ByteBuffer buffer = ByteBuffer.allocate(binary.length);
+        buffer.order(ByteOrder.LITTLE_ENDIAN);
+        buffer.put(binary);
+        String spirvTempDirectory = createSPIRVTempDirectoryName();
+        Path path = Paths.get(spirvTempDirectory);
+        try {
+            Files.createDirectories(path);
+        } catch (IOException e) {
+            throw new TornadoBailoutRuntimeException("Error - Exception when creating the temp directory for SPIR-V");
+        }
+        long timeStamp = System.nanoTime();
+        String pathSeparator = FileSystems.getDefault().getSeparator();
+        String spirvFile = spirvTempDirectory + pathSeparator + timeStamp + "-" + id + entryPoint + ".spv";
+        if (TornadoOptions.DEBUG) {
+            System.out.println("SPIR-V Binary File: " + spirvFile);
+        }
+
+        writeBufferToFile(buffer, spirvFile);
+        return installSPIRVBinary(meta, id, entryPoint, spirvFile);
+    }
+
     public abstract SPIRVInstalledCode installSPIRVBinary(TaskMetaData meta, String id, String entryPoint, String pathToFile);
-
-    public abstract SPIRVInstalledCode installSPIRVBinary(TaskMetaData meta, String id, String entryPoint, byte[] binary);
-
 }
